+++ mobile/android/android-components/docs/changelog.md
@@ -74,6 +74,10 @@ permalink: /changelog/
 * **browser-session**
   * Added `Session.hasParentSession` to indicate whether a `Session` was opened from a parent `Session` such as opening a new tab from a link context menu ("Open in new tab").
 
+* **feature-app-links**
+  * Add a flag to allow the app to not detect an external app if the user has told android to use the browser as default.
+  * Turn off interception of web links. 
+
 # 0.55.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.54.0...v0.55.0)